main/xiaozhi-server/requirements.txt
pyyml==0.0.2
torch==2.2.2
silero_vad==6.0.0
websockets==14.2
opuslib_next==1.1.2
numpy==1.26.4
pydub==0.25.1
funasr==1.2.3
torchaudio==2.2.2
openai==1.107.0
google-generativeai==0.8.5
edge_tts==7.0.0
httpx==0.27.2
aiohttp==3.12.15
aiohttp_cors==0.7.0
ormsgpack==1.7.0
ruamel.yaml==0.18.15
loguru==0.7.3
requests==2.32.5
cozepy==0.19.0
mem0ai==0.1.62
bs4==0.0.2
modelscope==1.23.2
sherpa_onnx==1.12.19
mcp==1.13.1
cnlunar==0.2.0
PySocks==1.7.1
dashscope==1.23.1
baidu-aip==4.16.13
chardet==5.2.0
aioconsole==0.8.1
markitdown==0.1.3
mcp-proxy==0.8.2
PyJWT==2.8.0
psutil==7.0.0
portalocker==3.2.0
Jinja2==3.1.6
vosk==0.3.45
google-cloud-firestore==2.17.0
elevenlabs==2.23.0
paho-mqtt==2.1.0
pytest==9.0.1
dateparser==1.2.1
